backend/requirements.txt
openai==0.27.0
fastapi==0.95.0
uvicorn==0.21.0
PyMuPDF==1.22.3
python-dotenv==1.0.0
python-multipart==0.0.6  # 追加
